decrease duplications
