+++ src/commands/Rooms/Create.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const Command = require('../../Command.js');
+const { getUsersForCall } = require('../../CommonFunctions');
 
 /**
  * Array of useable channel values
@@ -44,36 +45,6 @@ const lockedRegex = new RegExp('--locked', 'ig');
  */
 const textRegex = new RegExp('--text', 'ig');
 
-/**
- * Gets the list of users from the mentions in the call
- * @param {Message} message Channel message
- * @returns {Array.<User>} Array of users to send message
- */
-const getUsersForCall = (message) => {
-  const users = [];
-  if (message.mentions.roles) {
-    message.mentions.roles.forEach(role =>
-      role.members.forEach(member =>
-        users.push(member.user)));
-  }
-  if (message.mentions.users) {
-    message.mentions.users.forEach((user) => {
-      if (users.indexOf(user) === -1) {
-        users.push(user);
-      }
-    });
-  }
-  let authorIncluded = false;
-  users.forEach((user) => {
-    if (user.id === message.author.id) {
-      authorIncluded = true;
-    }
-  });
-  if (!authorIncluded) {
-    users.push(message.author);
-  }
-  return users;
-};
 
 /**
  * Create temporary voice/text channels (can be expanded in the future)
